+++ AzureCommunicationUI/AzureCommunicationUI/CallComposite.swift
@@ -44,7 +44,7 @@ public class CallComposite {
         dependencyContainer.registerDependencies(callConfiguration)
 
         setupColorTheming()
-        setupLocalization()
+        setupLocalization(with: dependencyContainer.resolve())
         let toolkitHostingController = makeToolkitHostingController(router: dependencyContainer.resolve(),
                                                                     logger: dependencyContainer.resolve(),
                                                                     viewFactory: dependencyContainer.resolve())
@@ -141,10 +141,9 @@ public class CallComposite {
         }
     }
 
-    private func setupLocalization() {
+    private func setupLocalization(with provider: LocalizationProvider) {
         if let localizationConfiguration = localizationConfiguration {
-            let localizationProvider = LocalizationProvider(logger: logger)
-            localizationProvider.applyLocalizationConfiguration(localizationConfiguration)
+            provider.apply(localeConfig: localizationConfiguration)
         }
     }