+++ AzureCommunicationUI/AzureCommunicationUI/Presentation/Style/LocalizationProvider.swift
@@ -6,67 +6,78 @@
 import Foundation
 import SwiftUI
 
-struct LocalizationProvider {
-    let logger: Logger?
-    static var locale: String = "en"
-    static var localizableFilename: String = ""
-    static var customTranslations: [String: String] = [:]
-    static var isRightToLeft: Bool = false
+protocol LocalizationProvider {
+    func apply(localeConfig: LocalizationConfiguration)
+    func getSupportedLanguages() -> [String]
+    func getLocalizedString(_ key: StringKey) -> String
+    func getLocalizedString(_ key: StringKey, _ args: CVarArg...) -> String
+}
 
-    static var supportedLocales: [String] = Bundle(for: CallComposite.self).localizations
+class AppLocalizationProvider: LocalizationProvider {
+    private let logger: Logger
+    private var locale: String = "en"
+    private var localizableFilename: String = ""
+    private var customTranslations: [String: String] = [:]
+    private var isRightToLeft: Bool = false
 
-    static var layoutDirection: LayoutDirection {
+    var supportedLocales: [String] = Bundle(for: CallComposite.self).localizations
+
+    init(logger: Logger) {
+        self.logger = logger
+    }
+
+    var layoutDirection: LayoutDirection {
         isRightToLeft ? .rightToLeft : .leftToRight
     }
-    static var layoutDirectionUIKit: UISemanticContentAttribute {
+    var layoutDirectionUIKit: UISemanticContentAttribute {
         isRightToLeft ? .forceRightToLeft : .forceLeftToRight
     }
 
-    func applyLocalizationConfiguration(_ localeConfig: LocalizationConfiguration) {
-        if !LocalizationProvider.supportedLocales.contains(localeConfig.locale) {
+    func apply(localeConfig: LocalizationConfiguration) {
+        if !supportedLocales.contains(localeConfig.locale) {
             let warningMessage = "Locale not supported by default for " +
             "`\(localeConfig.locale)`, if string for AzureCommunicationUI " +
             "localization keys not provided in custom Localizable.strings " +
             "or customString, strings will default to `en`"
-            logger?.warning(warningMessage)
+            logger.warning(warningMessage)
         }
 
-        LocalizationProvider.locale = localeConfig.locale
-        LocalizationProvider.localizableFilename = localeConfig.localizableFilename
-        LocalizationProvider.customTranslations = localeConfig.customTranslations
-        LocalizationProvider.isRightToLeft = localeConfig.isRightToLeft
+        locale = localeConfig.locale
+        localizableFilename = localeConfig.localizableFilename
+        customTranslations = localeConfig.customTranslations
+        isRightToLeft = localeConfig.isRightToLeft
     }
 
-    static func getSupportedLanguages() -> [String] {
+    func getSupportedLanguages() -> [String] {
         return supportedLocales
     }
 
-    static func getLocalizedString(_ key: String) -> String {
-        if let customTranslation = customTranslations[key] {
+    func getLocalizedString(_ key: StringKey) -> String {
+        if let customTranslation = customTranslations[key.rawValue] {
             return customTranslation
         }
 
         if let path = Bundle.main
             .path(forResource: locale, ofType: "lproj"),
            let bundle = Bundle(path: path) {
-            let customLocalizableString = NSLocalizedString(key,
+            let customLocalizableString = NSLocalizedString(key.rawValue,
                                                             tableName: localizableFilename,
                                                             bundle: bundle,
                                                             value: "localize_key_not_found",
-                                                            comment: key)
+                                                            comment: key.rawValue)
             if customLocalizableString != "localize_key_not_found" {
                 return customLocalizableString
             }
         }
-        return getPredefinedLocalizedString(key)
+        return getPredefinedLocalizedString(key.rawValue)
     }
 
-    static func getLocalizedString(_ key: String, _ args: CVarArg...) -> String {
+    func getLocalizedString(_ key: StringKey, _ args: CVarArg...) -> String {
         let stringFormat = getLocalizedString(key)
         return String(format: stringFormat, arguments: args)
     }
 
-    private static func getPredefinedLocalizedString(_ key: String) -> String {
+    private func getPredefinedLocalizedString(_ key: String) -> String {
         if let path = Bundle(for: CallComposite.self)
             .path(forResource: locale, ofType: "lproj"),
            let bundle = Bundle(path: path) {