+++ AzureCommunicationUI/AzureCommunicationUITests/CompositeViewModelFactoryMocking.swift
@@ -21,7 +21,7 @@ class CompositeViewModelFactoryMocking: CompositeViewModelFactory {
     var previewAreaViewModel: PreviewAreaViewModel?
     var setupControlBarViewModel: SetupControlBarViewModel?
     var errorInfoViewModel: ErrorInfoViewModel?
-    var audioDeviceListViewModel: AudioDeviceListViewModel?
+    var audioDevicesListViewModel: AudioDevicesListViewModel?
     var primaryButtonViewModel: PrimaryButtonViewModel?
     var iconButtonViewModel: IconButtonViewModel?
     var setupViewModel: SetupViewModel?
@@ -39,13 +39,15 @@ class CompositeViewModelFactoryMocking: CompositeViewModelFactory {
     func getSetupViewModel() -> SetupViewModel {
         return setupViewModel ?? SetupViewModel(compositeViewModelFactory: self,
                                                 logger: logger,
-                                                store: store)
+                                                store: store,
+                                                localizationProvider: LocalizationProviderMocking())
     }
 
     func getCallingViewModel() -> CallingViewModel {
         return callingViewModel ?? CallingViewModel(compositeViewModelFactory: self,
                                                     logger: logger,
-                                                    store: store)
+                                                    store: store,
+                                                    localizationProvider: LocalizationProviderMocking())
     }
 
     func makeIconButtonViewModel(iconName: CompositeIcon,
@@ -88,10 +90,10 @@ class CompositeViewModelFactoryMocking: CompositeViewModelFactory {
                                                                 action: action)
     }
 
-    func makeAudioDeviceListViewModel(dispatchAction: @escaping ActionDispatch,
-                                      localUserState: LocalUserState) -> AudioDeviceListViewModel {
-        return audioDeviceListViewModel ?? AudioDeviceListViewModel(dispatchAction: dispatchAction,
-                                                                    localUserState: localUserState)
+    func makeAudioDevicesListViewModel(dispatchAction: @escaping ActionDispatch,
+                                       localUserState: LocalUserState) -> AudioDevicesListViewModel {
+        return audioDevicesListViewModel ?? AudioDevicesListViewModel(dispatchAction: dispatchAction,
+                                                                      localUserState: localUserState)
     }
 
     func makeErrorInfoViewModel() -> ErrorInfoViewModel {