+++ AzureCommunicationUI/AzureCommunicationUITests/Presentation/Setup/SetupViewModelTests.swift
@@ -195,7 +195,8 @@ extension SetupViewModelTests {
 
     func makeSUT() -> SetupViewModel {
         return SetupViewModel(compositeViewModelFactory: factoryMocking,
-                                        logger: logger,
-                                        store: storeFactory.store)
+                              logger: logger,
+                              store: storeFactory.store,
+                              localizationProvider: LocalizationProviderMocking())
     }
 }